Update: 104-fibonacci.c

+++ 0x02-functions_nested_loops/104-fibonacci.c
@@ -27,23 +27,23 @@ int main(void)
  */
 void fib(int limit)
 {
-	unsigned long int a = 1;
-	unsigned long int b = 2;
-	unsigned long int c = a + b;
+	int a = 1;
+	int b = 2;
+	int c = a + b;
 
 	int counter = 2;
 
-	printf("%lu, ", a);
-	printf("%lu, ", b);
+	printf("%d, ", a);
+	printf("%d, ", b);
 
 	while (counter < (limit - 1))
 	{
-		printf("%lu, ", c);
+		printf("%d, ", c);
 		a = b;
 		b = c;
 		c = a + b;
 
 		counter += 1;
 	}
-	printf("%lu\n", c);
+	printf("%d\n", c);
 }
